requirements.txt
fastapi==0.115.0
uvicorn==0.34.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
